% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmistData}
\alias{dmistData}
\title{DMIST data}
\format{An object of class \code{list} of length 6.}
\usage{
dmistData
}
\description{
DMIST performance data manually transcribed from
Pisano2005_NEJM_v353p1773 and its supplement.
}
\details{
The performance data is organized in a list with the following elements.
   \itemize{
     \item \code{nObs.FFDM} [num] The number of FFDM observations (from Supplemental Table 2)
     \item \code{nObs.SFM} [num] The number of SFM observations (from Supplemental Table 2)
     \item \code{nObs.total} [num] the total number of observations (from Supplemental Table 2)
     \item \code{caseDist} [list] This list contains two items (from Supplemental Table 2)
       \itemize{
         \item \code{desc} [char] A description of the data.
         \item \code{caseDist.df} [data frame] The distribution of BIRADS scores from DMIST. \cr
           Row 1 is the distribution of FFDM BIRADS scores among DMIST cancers. \cr
           Row 2 is the distribution of SFM BIRADS scores among DMIST cancers. \cr
           Row 3 is the distribution of FFDM BIRADS scores among all DMIST cases. \cr
           Row 4 is the distribution of SFM BIRADS scores among all DMIST cases.
       }
     \item \code{caseDist.denseBreasts} [data frame] Caption to Figure 1 and row "T stage" of Table 2.
       One row of cancer=n1 detected by FFDM and SFM, cancer=n1 detected by SFM but not FFDM,
       cancer=n1 detected by FFDM but not SFM, cancer=n1 not detected by SFM or FFDM, non-cancer=n2, and total.
     \item \code{performance} [list]
     \itemize{
       \item \code{desc} [char] A description of the data taken from Figure 1 and the text of
         Pisano2005_NEJM_v353p1773.pdf. SE of AUC is taken to be 1/4 of the confidence
         interval specified. AUC determined using seven-point malignancy scale.
         TPF,FPF are from Table 2 of Pisano2005_NEJM_v353p1773-Suppl.pdf.
         TPF,FPF are determined using BIRADS. Cancer determined at 455 days for AUC, TPF, and FPF.
       \item \code{FFDM} [list] A list of the diagnostic performance statistics:
         \code{AUC, AUCse, CIbot, CItop, TPF, TPFse, TNF, TNFse}
       \item \code{SFM} [list] A list of the diagnostic performance statistics:
         \code{AUC, AUCse, CIbot, CItop, TPF, TPFse, TNF, TNFse}
       \item \code{FFDMminusSFM} [list] A list of the diagnostic performance statistics:
         \code{AUC, AUCse, CIbot, CItop, p, TPF, TPFse, TPFCIbot, TPFCItop, TPFp,
           TNF, TNFse, TNFCIbot, TNFCItop, TNFp}
     }
   }
}
\keyword{datasets}
