% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doViperStudyDesign.R
\name{doViperStudyDesign}
\alias{doViperStudyDesign}
\title{doViperStudyDesign}
\usage{
doViperStudyDesign(rawData)
}
\arguments{
\item{rawData}{One of the VIPER sub-study data frames.}
}
\value{
A list object that contains several ways
     to summarize the distribution of cases in the study accoring to case type.
     \itemize{
       \item \code{nObs} [num] The number of observations in the sub-study.
       \item \code{readers} [char] The reader IDs in the sub-study.
       \item \code{nR} [num] The number of readers in the sub-study.
       \item \code{nC} [data frame] The number of cases in the sub-study by case type.
         One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.
       \item \code{perReader} [list] This list contains two data frames: FFDM and SFM.
         Each of these data frames includes the number of cases evaluated by each reader by case type.
         20 readers == 20 Rows of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.
       \item \code{perReader.summary} [list] This list contains two data frames: FFDM and SFM.
         Each of these data frames includes the total number of cases evaluated by all readers by case type.
         One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.
     }
}
\description{
Extract basic study design info
}
\examples{
df <- split(viperData::viperObs455, viperData::viperObs455$desc)
df.studyDesign <- doViperStudyDesign(df$screeningLowP)
}
