% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viperObservations}
\alias{viperObservations}
\alias{viperObs365}
\alias{viperObs455}
\title{viperObservations}
\format{An object of class \code{data.frame} with 20382 rows and 35 columns.}
\usage{
viperObs365

viperObs455
}
\description{
The VIPER data sets were created by aggregating all the raw reader data files from the
  FDA VIPER study: Validation of Imaging Premarket Evaluation and Regulation (VIPER).
  The data sets differ in the truth labels. See the documentation details.
}
\details{
A data frame with 36 variables: \cr
  \itemize{
    \item \code{CaseReadOrder} \cr
    \item \code{CaseID} \cr
    \item \code{Modality} \cr
    \item \code{ReaderID} \cr
    \item \code{Study} \cr
    \item \code{Track} \cr
    \item \code{ReadSetLetter} \cr
    \item \code{Ctype} \cr
    \item \code{Compare.CaseIDs} \cr
    \item \code{Initial.Comparison} \cr
    \item \code{Final.Comparison} \cr
    \item \code{case_id} \cr
    \item \code{redcap_event_name} \cr
    \item \code{readerID} \cr
    \item \code{form_start_datetime} \cr
    \item \code{modality} \cr
    \item \code{study} \cr
    \item \code{viewbox} \cr
    \item \code{num_img} \cr
    \item \code{recall} \cr
    \item \code{miss_img} \cr
    \item \code{score_start_datetime} \cr
    \item \code{abnormality_type} \cr
    \item \code{recall_scale} \cr
    \item \code{score_end_datetime} \cr
    \item \code{Group} \cr
    \item \code{norecall_scale} \cr
    \item \code{set_letter} \cr
    \item \code{track} \cr
    \item \code{quality} \cr
    \item \code{UID} \cr
    \item \code{score} \cr
    \item \code{desc} \cr
    \item \code{read_order} \cr
    \item \code{Comment} \cr
    \item \code{RecallRegion} \cr
    \item \code{count}
  }

  \strong{viperObs365}: This dataset is identical to \code{\link{viperObs455}} except for the truth labels.
  The truth labels for this dataset are based on cancer at 365 days.

  \strong{viperObs455}: This dataset is identical to \code{\link{viperObs365}} except for the truth labels.
  The truth labels for this dataset are based on cancer at 455 days.

  The images used in this study were selected from the DMIST images [Pisano2005_NEJM_v353p1773].
  The VIPER study concept 2011. Data collection September 2013 to August 2015.
  Please refer to journal paper for details.
}
\keyword{datasets}
