% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viperSummary}
\alias{viperSummary}
\alias{viperSummary365}
\alias{viperSummary455}
\title{viperSummaries}
\format{An object of class \code{list} of length 5.}
\usage{
viperSummary365

viperSummary455
}
\description{
The VIPER summary files contain summary statistics of the VIPER observations.
The iMRMC R package is used to estimate the area under the ROC curve (AUC),
the true positive fraction (sensitivity), and the true negative fraction (specificity).
The iMRMC package also estimates variances, standard errors, and confidence intervals
that account for the variability from readers and cases. You can refer to the R function
\code{doViperSummary.R} in this package for the code that produces the summary files.

The iMRMC R package can be found at
\itemize{
  \item{\url{https://github.com/DIDSR/iMRMC/releases},}
  \item{\url{https://github.com/DIDSR/iMRMC/tree/master/Rpackage}, and}
  \item{\url{https://cran.r-project.org/web/packages/iMRMC/index.html}.}
}
}
\details{
Each element of the list pertains to one of the VIPER sub-studies:
  \itemize{
    \item \code{screeningLowP} [list] Screening population, per reader prevalence ~ 10\%
    \item \code{screeningMedP} [list] Screening population, per reader prevalence ~ 30\%
    \item \code{screeningHighP} [list] Screening population, per reader prevalence ~ 50\%
    \item \code{challengeMedP} [list] Challenge population, per reader prevalence ~ 30\%
    \item \code{challengeHighP} [list] Challenge population, per reader prevalence ~ 50\%
  }

Each substudy list element is a list object of length 3.
  \itemize{
    \item \code{desc} [string] Sub-study label. This is the same as the sub-study list element name.
    \item \code{studyDesign} [list] This list object contains several ways
     to summarize the distribution of cases in the study accoring to case type.
     \itemize{
       \item \code{nObs} [num] The number of observations in the sub-study.
       \item \code{readers} [char] The reader IDs in the sub-study.
       \item \code{nR} [num] The number of readers in the sub-study.
       \item \code{nC} [data frame] The number of cases in the sub-study by case type.
         One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.
       \item \code{perReader} [list] This list contains two data frames: FFDM and SFM.
         Each of these data frames includes the number of cases evaluated by each reader by case type.
         20 readers == 20 Rows of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.
       \item \code{perReader.summary} [list] This list contains two data frames: FFDM and SFM.
         Each of these data frames includes the total number of cases evaluated by all readers by case type.
         One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.
     }
    \item \code{iMRMC} [list] This list object contains the MRMC analysis results for
     auc, tpf, and tnf.
     There are four groups of results: perReader, Ustat, MLEstat, ROC, and varDecomp
     The analysis results are produced by a command line version of iMRMC called by R:
     \url{https://github.com/DIDSR/iMRMC/releases},
     \url{https://github.com/DIDSR/iMRMC/tree/master/Rpackage},
     \url{https://cran.r-project.org/web/packages/iMRMC/index.html}
  }

  \strong{viperSummary365}: This summary is based on \code{\link{viperObs365}} the dataset
  based on cancer at 365 days.

  \strong{viperSummary455}: This summary is based on \code{\link{viperObs455}} the dataset
  based on cancer at 455 days.
}
\keyword{datasets}
